refactor: add 'utils' for methods and rename ButtonContext to HandlerInputContext
