+++ src/components/Button/index.jsx
@@ -1,16 +1,15 @@
 import React, { useContext } from 'react';
-import Context from 'context/ButtonContext';
+import Context from 'context/HandlerInputContext';
 
 import './styles.css';
 
 export default function Button({ digit, className }) {
-  const { keyword, handleChange, inputDisplay } = useContext(Context);
+  const { keyword, handleChange } = useContext(Context);
 
   const handleClick = (e) => {
     const newDigit = e.target.value;
     const beforeValue = keyword ?? 0;
     handleChange(beforeValue, newDigit);
-    inputDisplay;
   };
 
   return (
