+++ src/components/Display/index.jsx
@@ -1,6 +1,6 @@
 import React, { useEffect, useContext } from 'react';
 import './styles.css';
-import Context from 'context/ButtonContext';
+import Context from 'context/HandlerInputContext';
 
 export default function Display() {
   const { keyword, handleChange, inputDisplay } = useContext(Context);
@@ -8,11 +8,6 @@ export default function Display() {
   useEffect(() => {
     inputDisplay.current.focus();
   }, []);
-
-  // TODO: que los botones funcionen más un efecto de hover y presionado.
-  // TODO: al presionar los números desde el teclado, los botones se hagan un efecto de ser presionados
-  // TODO: sumar, restar, dividir y multiplicar. Quizás mostrar un pequeño íncono de lo que se está haciendo
-  // TODO: Para moviles, deshabilitar el input y que tenga que usar los botones de la calculador para ingresar
 
   return (
     <div className='calculator--display'>
@@ -27,6 +22,7 @@ export default function Display() {
         onChange={handleChange}
         value={keyword ?? 0}
         ref={inputDisplay}
+        autoComplete='off'
       />
     </div>
   );
